requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.3
boto3==1.41.2
botocore==1.41.2
brotli==1.2.0
celery==5.5.3
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.8
django-cors-headers==4.3.1
django-filter==25.2
django-ses==3.5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
fonttools==4.60.1
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
pydyf==0.11.0
PyJWT==2.8.0
pypdf==6.4.0
PyPDF2==3.0.1
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
